# swagger.yaml
openapi: 3.0.0
info:
  title: ToDo-API
  description: A simple todo-api.
  version: 0.1.9
servers:
  - url: http://localhost:3000
paths:
  /todos:
    get:
      summary: Returns all todos
      responses:
        "200":
          description: "Successful get"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
    post:
      summary: Create a new todo
      requestBody: 
        required: true
        content:
          application/json:
            schema: 
              type: object
              properties:
                task:
                  type: string
                  example: "New task"
      responses: 
        "201":
          description: "Successfully created"
          content: 
            application/json:
              schema: 
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  completed:
                    type: boolean
                    example: false
  /todos/{id}:
    get:
      summary: Get a todo by ID
      parameters: 
        - name: id
          in: path
          required: true
          description: The ID of the todo to retrieve
          schema: 
            type: integer
      responses: 
        "200":
          description: "Successful get"
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  id:
                    type: integer
                    example: 1
                  task:
                    type: string
                    example: "Sample task"
                  completed:
                    type: boolean
                    example: false
        "404":
          description: "ID not found"
    put: 
      summary: Update a todo by ID
      parameters: 
        - name: id
          in: path
          required: true
          description: The ID of the todo to update
          schema: 
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema: 
              type: object
              properties:
                task:
                  type: string
                  example: "Updated task"
      responses:
        "200":
          description: "Successfully updated"
          content: 
            application/json:
              schema: 
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  completed:
                    type: boolean
                    example: false
        "404":
          description: "Task not found"
    delete:
      summary: Delete a todo by ID
      parameters:
        - name: id
          in: path
          required: true
          description: "The ID of the todo to delete"
          schema: 
            type: integer
      responses:
        "204":
          description: "Successfully deleted"
        "404":
          description: "Task not found!"